Moved simulation readers into a subdirectory. Fixed some paths to the files
