+++ nfwfitter/readers/readAnalytic.py
